+++ Editor/ProjectAuditorParams.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Unity.ProjectAuditor.Editor.AssemblyUtils;
 using UnityEditor;
 
 namespace Unity.ProjectAuditor.Editor
@@ -8,6 +9,8 @@ namespace Unity.ProjectAuditor.Editor
     {
         public IssueCategory[] categories;
         public BuildTarget platform;
+        public string[] assemblyNames;
+        public CodeOptimization codeOptimization;
 
         public Action<IReadOnlyCollection<ProjectIssue>> onAuditAsyncUpdate;
         public Action<ProjectReport> onAuditAsyncComplete;
@@ -15,12 +18,15 @@ namespace Unity.ProjectAuditor.Editor
         public ProjectAuditorParams()
         {
             platform = EditorUserBuildSettings.activeBuildTarget;
+            codeOptimization = CodeOptimization.Release;
         }
 
         public ProjectAuditorParams(ProjectAuditorParams original)
         {
             categories = original.categories;
             platform = original.platform;
+            assemblyNames = original.assemblyNames;
+            codeOptimization = original.codeOptimization;
 
             onAuditAsyncUpdate = original.onAuditAsyncUpdate;
             onAuditAsyncComplete = original.onAuditAsyncComplete;