+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Added *ProjectAuditorParams* option for compiling selected assemblies
 * Renamed asynchronous *ProjectAuditor.Audit* to *AuditAsync*
 
 ## [0.8.0-preview] - 2022-05-20